Refactor: post 관련 component 이름 수정
- 폴더명에 Post 삭제
