+++ frontend/src/components/Post/Content/style.ts
