+++ frontend/src/components/Post/Image/index.tsx
