+++ frontend/src/components/Post/Image/style.ts
