+++ frontend/src/components/Post/LikeRow/style.ts
